Move formatting into a formatter, thinking ahead
